requirements.txt
asgiref==3.6.0
dj-static==0.0.6
Django==4.1.5
django-stdimage==6.0.1
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow
psycopg2-binary==2.9.5
sqlparse==0.4.3
static3==0.7.0
